Cleaning up errors
